Python/web-scraper-to-sqlite/requirements.txt
requests
beautifulsoup4
pandas
numpy
pyyaml
lxml
aiohttp
pytest
black
click
python-dateutil
pytz
python-dotenv
tenacity
backoff
ratelimit
structlog
sqlalchemy
alembic
pydantic
rich
matplotlib
seaborn
plotly
sphinx
sphinx-rtd-theme